topic card snapshots
