+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.TopicCardViewControllerSnapshotTests/test_loadInNavigationController_dark_rendersCorrectly@3x.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:eb4cc574cdf0ab761d3a4a4023ee04ab40161eb8a3aff591db8429af6da239d1
+size 77992
