+++ Tests/govuk_ios/govuk_ios_snapshot_tests/ReferenceImages_64/govuk_ios_snapshot_tests.TopicCardViewControllerSnapshotTests/test_loadInNavigationController_light_rendersCorrectly@3x.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:70e2c68e4c133584517edbe7b5fe786896baf40bc7414c32e1861ce0a7eb2cf8
+size 70289
